pyproject.toml
[tool.poetry]
name = "GF"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
gunicorn = "20.1.0"
psycopg2-binary = "2.9.1"
redis = "3.5.3"
python = "^3.9"
Django = "3.2.9"
djangorestframework = "3.12.4"
django-allauth = "0.45.0"
django-filter = "2.4.0"
django-structlog = "2.1.3"
whitenoise = "5.3.0"
celery = "5.1.2"
flower = "1.0.0"
user-agents = "2.2.0"
debugpy = "1.5.1"
django-debug-toolbar = "3.2.2"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
pytest-django = "4.4.0"
pytest-factoryboy = "2.1.0"
factory-boy = "3.2.1"
mypy = "0.910"
bandit = "1.7.0"
pylint = "2.11.1"
pylint-django = "2.4.4"
flake8 = "4.0.1"
black = "21.10b0"
pre-commit = "2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
